AGENTIC-AI-CLASSES/class6-assignment/pyproject.toml
[project]
name = "class6-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.6.5",
    "openai-agents>=0.2.4",
]
